KD8ZRC.Mapview.Utility.Websocket

Description

Provides websocket callbacks for mapview 3.0+

Documentation

type Client = (String, Chan ByteString)

type ServerState = [Client]

data WebsocketServer

Constructors

WebsocketServer (MVar ServerState)

newtype WebsocketOnConnectCallback

These callbacks get called when a client connects to the websocket. They can be used to send a welcome message of sorts, or more practically, a history of previous downlink packet information.

Constructors

WebsocketOnConnectCallback (Client -> IO ())

callOnConnectCallback :: WebsocketOnConnectCallback -> Client -> IO ()

A simple wrapper to call a WebsocketOnConnectCallback with a given Client.

initWebsocketServer :: Chan ByteString -> String -> Int -> [WebsocketOnConnectCallback] -> IO ()

To make use of mapview-websocket in your MapView application, start off by initializing a new 'Chan.Chan T.Text'. To do this, in your main' function, write rawChan <- Chan.newChan. This should get passed to your MapviewConfig and sent to writeChanRaw in your _mvPacketLineCallback list.

Now you can start a WebsocketServer. Still in main, fork a WebsocketServer by doing (something like) _ <- forkIO $ initWebsocketServer rawChan "127.0.0.1" 8181, changing the hostname and port number to bind to, as necessary.

As MapView updates rawChan (via the writeChanRaw callback), mapview-websocket will see the updates and broadcast them to clients as soon as it can.

addClient :: Client -> ServerState -> ServerState

removeClient :: Client -> ServerState -> ServerState

broadcast :: Chan ByteString -> MVar ServerState -> IO ()

application :: MVar ServerState -> [WebsocketOnConnectCallback] -> ServerApp